+++ examples/pong/pong.rs
@@ -7,8 +7,9 @@ use amethyst::{
     ecs::prelude::World,
     prelude::*,
     renderer::{
-        Camera, MaterialTextureSet, PngFormat, Projection, Sprite, SpriteRender, SpriteSheet,
-        SpriteSheetHandle, Texture, TextureCoordinates, WindowMessages,
+        Camera, FilterMethod, MaterialTextureSet, PngFormat, Projection, Sprite, SpriteRender,
+        SpriteSheet, SpriteSheetHandle, Texture, TextureCoordinates, TextureMetadata,
+        WindowMessages,
     },
     ui::{Anchor, TtfFormat, UiText, UiTransform},
 };
@@ -52,7 +53,7 @@ fn load_sprite_sheet(world: &mut World) -> SpriteSheetHandle {
         loader.load(
             "texture/pong_spritesheet.png",
             PngFormat,
-            Default::default(),
+            TextureMetadata::srgb().with_filter(FilterMethod::Scale),
             (),
             &texture_storage,
         )